fix: refactor analytics logic
